# Conflicts:
#	invoicing/src/main/java/com/kynsoft/finamer/invoicing/application/command/manageInvoice/send/SendInvoiceCommandHandler.java

+++ creditCard/src/main/java/com/kynsoft/finamer/creditcard/domain/dto/PaymentRequestDto.java
@@ -13,5 +13,5 @@ public class PaymentRequestDto {
     private UUID merchantId;
     private String amount;
     private String orderNumber;
-    private int transactionId;
+    private long transactionId;
 }
